chore(compiler): move addLeadingJSDocComment to compiler utils
